Move bundle to its own folder
